Move elementary test of JW mapping to a new file
